Add security to views for products admin

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -80,11 +81,16 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     View to add product to website
     Code taken from Boutique Ado CI tutorial
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do this action')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -105,11 +111,16 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     View to edit product to website
     Code taken from Boutique Ado CI tutorial
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do this action')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -134,11 +145,16 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     View to delete product from website
     Code taken from Boutique Ado CI tutorial
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do this action')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
